fix(runtime): pair the new prototype reads with the handle-band predicate; split two files at the cap
The raw-pointer branch used a 0x10000 floor where HANDLE_BAND_MAX is 0x100000,
so it admitted handle-band values and handed them to a dereference; both new
is_valid_obj_ptr guards were unpaired (#6279). Splits array/indexing.rs and
get_field_by_name_tail.rs, which crossed the 2000-line cap.
