+++ crates/perry-runtime/src/object/field_get_set.rs
@@ -197,6 +197,9 @@ pub(crate) fn is_fetch_subclass_body_method(name: &[u8]) -> bool {
 // ── Topical sub-modules (issue #1103: keep every file < 2000 lines) ──
 mod accessors;
 pub(crate) use accessors::scan_accessor_receiver_override_root_mut;
+/// #9192 array prototype/constructor slots, split out of
+/// `get_field_by_name_tail.rs` for the 2000-line cap.
+mod array_retargeted_proto;
 mod buffer_own_prop;
 mod class_object_props;
 mod crypto_key;
